requirements (1).txt
# Function dependencies, for example:
# package>=version
pysftp
google-cloud-storage
google-cloud-storage
